Change BalanceSheet amounts from int to float64
The amounts returned are typically integers. However, some of the
amounts returned for UBNT's balance sheet are floats; therefore, I
changed all amounts to float64 from int.

+++ balancesheet.go
@@ -14,35 +14,37 @@ type BalanceSheets struct {
 	Statements []BalanceSheet `json:"balancesheet"`
 }
 
-// BalanceSheet models one balance sheet statement.
+// BalanceSheet models one balance sheet statement. Normally the amounts
+// returned are integers, although the currentCash for UBNT returned is a
+// float; therefore, these are all floats.
 type BalanceSheet struct {
-	ReportDate              Date `json:"reportDate"`
-	CurrentCash             int  `json:"currentCash"`
-	ShortTermInvestments    int  `json:"shortTermInvestments"`
-	Receivables             int  `json:"receivables"`
-	Inventory               int  `json:"inventory"`
-	OtherCurrentAssets      int  `json:"otherCurrentAssets"`
-	CurrentAssets           int  `json:"currentAssets"`
-	LongTermInvestments     int  `json:"longTermInvestments"`
-	PropertyPlanetEquipment int  `json:"propertyPlantEquipment"`
-	Goodwill                int  `json:"goodwill"`
-	IntangibleAssets        int  `json:"intangibleAssets"`
-	OtherAssets             int  `json:"otherAssets"`
-	TotalAssets             int  `json:"totalAssets"`
-	AccountsPayable         int  `json:"accountsPayable"`
-	CurrentLongTermDebt     int  `json:"currentLongTermDebt"`
-	OtherCurrentLiabilities int  `json:"otherCurrentLiabilities"`
-	TotalCurrentLiabilities int  `json:"totalCurrentLiabilities"`
-	LongTermDebt            int  `json:"longTermDebt"`
-	OtherLiablities         int  `json:"otherLiabilities"`
-	MinorityInterest        int  `json:"minorityInterest"`
-	TotalLiabilities        int  `json:"totalLiabilities"`
-	CommonStock             int  `json:"commonStock"`
-	RetainedEarnings        int  `json:"retainedEarnings"`
-	TreasuryStock           int  `json:"treasuryStock"`
-	CapitalSurplus          int  `json:"capitalSurplus"`
-	ShareholderEquity       int  `json:"shareholderEquity"`
-	NetTangibleAssets       int  `json:"netTangibleAssets"`
+	ReportDate              Date    `json:"reportDate"`
+	CurrentCash             float64 `json:"currentCash"`
+	ShortTermInvestments    float64 `json:"shortTermInvestments"`
+	Receivables             float64 `json:"receivables"`
+	Inventory               float64 `json:"inventory"`
+	OtherCurrentAssets      float64 `json:"otherCurrentAssets"`
+	CurrentAssets           float64 `json:"currentAssets"`
+	LongTermInvestments     float64 `json:"longTermInvestments"`
+	PropertyPlanetEquipment float64 `json:"propertyPlantEquipment"`
+	Goodwill                float64 `json:"goodwill"`
+	IntangibleAssets        float64 `json:"intangibleAssets"`
+	OtherAssets             float64 `json:"otherAssets"`
+	TotalAssets             float64 `json:"totalAssets"`
+	AccountsPayable         float64 `json:"accountsPayable"`
+	CurrentLongTermDebt     float64 `json:"currentLongTermDebt"`
+	OtherCurrentLiabilities float64 `json:"otherCurrentLiabilities"`
+	TotalCurrentLiabilities float64 `json:"totalCurrentLiabilities"`
+	LongTermDebt            float64 `json:"longTermDebt"`
+	OtherLiablities         float64 `json:"otherLiabilities"`
+	MinorityInterest        float64 `json:"minorityInterest"`
+	TotalLiabilities        float64 `json:"totalLiabilities"`
+	CommonStock             float64 `json:"commonStock"`
+	RetainedEarnings        float64 `json:"retainedEarnings"`
+	TreasuryStock           float64 `json:"treasuryStock"`
+	CapitalSurplus          float64 `json:"capitalSurplus"`
+	ShareholderEquity       float64 `json:"shareholderEquity"`
+	NetTangibleAssets       float64 `json:"netTangibleAssets"`
 }
 
 // AnnualBalanceSheets returns the annual balance sheets from the IEX Cloud
